MS2/requirements.txt
psycopg2-binary
flask
pika